Django/SW-API/apps/posts/views/views.py
from django.http import HttpResponse
from django.shortcuts import render
from rest_framework.decorators import api_view
from rest_framework.response import Response

# Create your views here.

@api_view(['GET', 'POST'])
def posts_api(request):
    return Response('Hello, World!')